# filters: optimize cel expression context with constant-time lookups