--- a/test/extensions/filters/common/expr/expr_context_speed_test.cc
+++ b/test/extensions/filters/common/expr/expr_context_speed_test.cc
@@ -167,8 +167,5 @@ BENCHMARK(bmResponseAttributes)
     ->RangeMultiplier(100)
     ->Range(100, 1000000);
 
-BENCHMARK(bmFilterState)
-    ->Unit(::benchmark::kMicrosecond)
-    ->RangeMultiplier(100)
-    ->Range(10, 100000);
+BENCHMARK(bmFilterState)->Unit(::benchmark::kMicrosecond)->RangeMultiplier(100)->Range(10, 100000);
 } // namespace Envoy