## Description

This PR optimizes the CEL expression context implementation by replacing `if-else` chains with hash maps for `O(1)` lookups and improving memory management. This would enhance the performance for header and filter state access patterns while maintaining the same functionality.

This change introduces constant-time lookups using `absl::flat_hash_map` and improves the overall design through static initialization of lookup maps for better performance.

## Benchmarks

### Before

```
Run on (32 X 3364.47 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 2048 KiB (x16)
  L3 Unified 107520 KiB (x1)
Load Average: 9.07, 16.88, 15.22
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
bmRequestAttributes/100             0.179 us        0.179 us      3943200
bmRequestAttributes/10000           0.178 us        0.178 us      3768352
bmRequestAttributes/1000000         0.179 us        0.179 us      3882288
bmResponseAttributes/100            0.187 us        0.187 us      3734635
bmResponseAttributes/10000          0.180 us        0.180 us      3778717
bmResponseAttributes/1000000        0.196 us        0.196 us      3668511
bmFilterState/1000                   51.6 us         51.6 us        10000
bmFilterState/10000                   692 us          692 us         1025
bmFilterState/1000000              262413 us       262406 us            3
bmFilterState/10000000            3887577 us      3887332 us            1
bmConnectionAttributes/100          0.465 us        0.465 us      1521335
bmConnectionAttributes/10000        0.467 us        0.467 us      1500911
bmConnectionAttributes/1000000      0.468 us        0.468 us      1521145
```

### After

```
Run on (32 X 3269.07 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x16)
  L1 Instruction 32 KiB (x16)
  L2 Unified 2048 KiB (x16)
  L3 Unified 107520 KiB (x1)
Load Average: 15.79, 17.71, 15.83
-------------------------------------------------------------------------
Benchmark                               Time             CPU   Iterations
-------------------------------------------------------------------------
bmRequestAttributes/100             0.166 us        0.166 us      4024447
bmRequestAttributes/10000           0.167 us        0.167 us      4027806
bmRequestAttributes/1000000         0.171 us        0.171 us      4217833
bmResponseAttributes/100            0.174 us        0.174 us      3933317
bmResponseAttributes/10000          0.169 us        0.169 us      4106892
bmResponseAttributes/1000000        0.173 us        0.173 us      4110847
bmFilterState/1000                   52.1 us         52.1 us        14034
bmFilterState/10000                   655 us          655 us         1078
bmFilterState/1000000              251507 us       251496 us            3
bmFilterState/10000000            3806409 us      3806192 us            1
bmConnectionAttributes/100          0.435 us        0.435 us      1613013
bmConnectionAttributes/10000        0.423 us        0.423 us      1674895
bmConnectionAttributes/1000000      0.411 us        0.411 us      1729797
```

---

**Commit Message:** filters: optimize cel expression context with constant-time lookups
**Additional Description:** This commit replaces linear if-else chains with flat_hash_maps for property lookups in the CEL expression context. It also improves memory management and provides better encapsulation of implementation details.

**Risk Level:** Low
- No functional changes
- Only performance optimizations
- Existing test coverage maintained

**Testing:**
- All existing tests pass
- No behavioral changes

**Docs Changes:** N/A - internal implementation change only
**Release Notes:** N/A