@wbpcode 2024-11-09
Thanks. This is great. Any performance improvements are welcome if it does't break the security and our style.

But could you also add a benchmark test to prove it works as our expection? Thanks.

/wait