@agrawroh 2024-11-09
> Thanks. This is great. Any performance improvements are welcome if it does't break the security and our style.
> 
> 
> 
> But could you also add a benchmark test to prove it works as our expection? Thanks.
> 
> 
> 
> /wait

@wbpcode Thanks for taking a look. I can certainly add a benchmarking test. This would be my first time. If possible, could you please share some pointers on how to write such a test?