@wbpcode 2024-11-09
> > Thanks. This is great. Any performance improvements are welcome if it does't break the security and our style.
> > But could you also add a benchmark test to prove it works as our expection? Thanks.
> > /wait
> 
> @wbpcode Thanks for taking a look. I can certainly add a benchmarking test. This would be my first time. If possible, could you please share some pointers on how to write such a test?

Every test is different. But you can find lots test files that end with `benchmark` or contain `speed`. There are lots of test cases. Besically, you only need to define a static function to run the code you want to test. And then the benchmark macro could help you to resolve other things.