@wbpcode 2024-11-12
Seems the performance improvement is trival. But I also agree the new look up table pattern is better. So, will take a look tomorrow.