@kyessenov 2024-11-12
LGTM apart from a change in filter_state lookup logic. I agree, this is better, and would allow us to implement `Keys()` quickly in a followup, even though the performance benefit is not significant.