@kyessenov 2024-11-15
@agrawroh What's up with filter state benchmark:

```
bmFilterState/10000000            3806409 us      3806192 us            1
```

Is it a benchmark problem or some implementation difference?